requirements.txt
Flask
Flask-HTTPAuth
RouterOS-api
requests
python-dotenv
gevent
git+https://github.com/esoadamo/Python-SQLiteDB.git
